server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.2
orjson==3.10.6
numpy==2.0.1
scikit-learn==1.5.1
pymupdf==1.24.7
openai==1.40.6
python-dotenv==1.0.1
